+++ apis/r/R/SOMAArrayBase.R
@@ -30,11 +30,12 @@ SOMAArrayBase <- R6::R6Class(
     },
 
     write_object_type_metadata = function() {
-      private$check_open_for_write()
+      #private$check_open_for_write()
 
       meta <- list()
       meta[[SOMA_OBJECT_TYPE_METADATA_KEY]] <- self$class()
       meta[[SOMA_ENCODING_VERSION_METADATA_KEY]] <- SOMA_ENCODING_VERSION
+      spdl::debug("[SOMAArrayBase::write_object_metadata] calling set metadata")
       self$set_metadata(meta)
     }
   )