+++ apis/r/R/TileDBObject.R
@@ -39,11 +39,16 @@ TileDBObject <- R6::R6Class(
       private$.tiledb_ctx <- self$tiledbsoma_ctx$context()
 
       if (!is.null(tiledb_timestamp)) {
-        stopifnot("'tiledb_timestamp' must be a POSIXct datetime object" = inherits(tiledb_timestamp, "POSIXct"))
-        private$tiledb_timestamp <- tiledb_timestamp
+        stopifnot(
+            "'tiledb_timestamp' must be a single POSIXct datetime object" = inherits(tiledb_timestamp, "POSIXct") &&
+              length(tiledb_timestamp) == 1L &&
+              !is.na(tiledb_timestamp)
+        )
+        private$.tiledb_timestamp <- tiledb_timestamp
       }
 
-      spdl::debug("[TileDBObject] initialize {} with '{}'", self$class(), self$uri)
+      spdl::debug("[TileDBObject] initialize {} with '{}' at ({})", self$class(), self$uri,
+                  self$tiledb_timestamp %||% "now")
     },
 
     #' @description Print the name of the R6 class.
@@ -82,13 +87,18 @@ TileDBObject <- R6::R6Class(
     #'  \item \dQuote{\code{READ}}
     #'  \item \dQuote{\code{WRITE}}
     #' }
+    #' @param tiledb_timestamp Optional Datetime (POSIXct) with TileDB timestamp
     #'
     #' @return Invisibly returns \code{self} opened in \code{mode}
     #'
-    reopen = function(mode) {
+    reopen = function(mode, tiledb_timestamp = NULL) {
       mode <- match.arg(mode, choices = c('READ', 'WRITE'))
+      stopifnot(
+        "'tiledb_timestamp' must be a POSIXct datetime object" = is.null(tiledb_timestamp) ||
+          (inherits(tiledb_timestamp, what = "POSIXct") && length(tiledb_timestamp) == 1L && !is.na(tiledb_timestamp))
+      )
       self$close()
-      private$tiledb_timestamp <- NULL
+      private$.tiledb_timestamp <- tiledb_timestamp
       self$open(mode, internal_use_only = 'allowed_use')
       return(invisible(self))
     },
@@ -130,11 +140,33 @@ TileDBObject <- R6::R6Class(
       }
       return(private$.tiledbsoma_ctx)
     },
+    #' @field tiledb_timestamp Time that object was opened at
+    #'
+    tiledb_timestamp = function(value) {
+      if (!missing(value)) {
+        private$.read_only_error("tiledb_timestamp")
+      }
+      return(private$.tiledb_timestamp)
+    },
     #' @field uri
     #' The URI of the TileDB object.
     uri = function(value) {
       if (missing(value)) return(private$tiledb_uri$uri)
       stop(sprintf("'%s' is a read-only field.", "uri"), call. = FALSE)
+    },
+    #' @field .tiledb_timestamp_range Time range for libtiledbsoma
+    #'
+    .tiledb_timestamp_range = function(value) {
+      if (!missing(value)) {
+        private$.read_only_error("tiledb_timestamp_range")
+      }
+      if (is.null(self$tiledb_timestamp)) {
+        return(NULL)
+      }
+      return(c(
+        as.POSIXct(0, tz = 'UTC', origin = '1970-01-01'),
+        self$tiledb_timestamp
+      ))
     }
   ),
 
@@ -162,7 +194,7 @@ TileDBObject <- R6::R6Class(
 
     # Opener-supplied POSIXct timestamp, if any. TileDBArray and TileDBGroup are each responsible
     # for making this effective, since the methods differ slightly.
-    tiledb_timestamp = NULL,
+    .tiledb_timestamp = NULL,
 
     # Internal context
     .tiledbsoma_ctx = NULL,